decompose expr_statement
